+++ src/launcher/java/org/truffleruby/launcher/RubyLauncher.java
@@ -228,8 +228,12 @@ public class RubyLauncher extends AbstractLanguageLauncher {
             Metrics.printTime("after-run");
             return exitCode;
         } catch (PolyglotException e) {
-            System.err.println(
-                    "truffleruby: an exception escaped out of the interpreter - this is an implementation bug");
+            if (e.isHostException()) {
+                System.err.println("truffleruby: a host exception reached the top level:");
+            } else {
+                System.err.println(
+                        "truffleruby: an exception escaped out of the interpreter - this is an implementation bug");
+            }
             e.printStackTrace();
             return 1;
         }