+++ src/main/java/org/truffleruby/language/backtrace/BacktraceFormatter.java
@@ -391,6 +391,11 @@ public class BacktraceFormatter {
 
     @TruffleBoundary
     public static void printInternalError(RubyContext context, Throwable throwable, String from) {
+        if (context.getEnv().isHostException(throwable)) {
+            // rethrow host exceptions to get the interleaved host and guest stacktrace of PolyglotException
+            throw ExceptionOperations.rethrow(throwable);
+        }
+
         final PrintStream stream = BacktraceFormatter.printStreamFor(context.getEnv().err());
         final BacktraceFormatter formatter = context.getDefaultBacktraceFormatter();
         stream.println();