+++ CHANGELOG.md
@@ -7,6 +7,7 @@ New features:
 Bug fixes:
 
 * `Range#to_a` wasn't working for `long` ranges (#2198, @tomstuart and @LillianZ).
+* Show the interleaved host and guest stacktrace for host exceptions (#2226).
 
 Compatibility: